Functions

This page briefly describes the functions exported by the package.

The function names in this module are in lowercase letters and sometimes underscores are used for longer function names. Most functions require keyword arguments, i.e. tcia_bodyparts(collection = collection_name) will work but tcia_bodyparts(collection_name) will not.

Index

CancerImagingArchive.dataframe_to_csv

CancerImagingArchive.dictionary_to_json

CancerImagingArchive.remove_empty!

CancerImagingArchive.tcia_bodyparts

CancerImagingArchive.tcia_collections

CancerImagingArchive.tcia_images

CancerImagingArchive.tcia_manufacturers

CancerImagingArchive.tcia_modalities

CancerImagingArchive.tcia_newstudies

CancerImagingArchive.tcia_patients

CancerImagingArchive.tcia_patients_by_modality

CancerImagingArchive.tcia_series

CancerImagingArchive.tcia_series_size

CancerImagingArchive.tcia_single_image

CancerImagingArchive.tcia_sop

CancerImagingArchive.tcia_studies

Public functions

CancerImagingArchive.dataframe_to_csv — Method

dataframe_to_csv(; dataframe, file)

Writes the information in a DataFrame object (dataframe) into a csv file (file).

CancerImagingArchive.dictionary_to_json — Method

dictionary_to_json(; dictionary, file)

Writes the information in a Dictionary Array (dictionary) into a json file (file).

CancerImagingArchive.tcia_bodyparts — Method

tcia_bodyparts(; collection, modality, format = "csv")

Returns the body parts examined in a given collection and/or by a given modality.

CancerImagingArchive.tcia_collections — Method

tcia_collections(; format = "csv")

Provides names of all the collections on TCIA.

CancerImagingArchive.tcia_images — Method

tcia_images(; series, file)

Downloads the images for the given SeriesInstanceUID series as the given zip-file file.

CancerImagingArchive.tcia_manufacturers — Method

tcia_manufacturers(; collection, modality, bodypart, format = "csv")

Returns the hardware manufacturers for a given collection and/or modality and/or bodypart.

CancerImagingArchive.tcia_modalities — Method

tcia_modalities(; collection, bodypart, format = "csv")

Returns the modalities used in a given collection and/or for a given bodypart.

CancerImagingArchive.tcia_newstudies — Method

tcia_newstudies(; date, collection, patient, format = "csv")

Returns new studies for a given collection that were added after a given date formatted as YYYY-MM-DD. The patient ID can be optionally given.

CancerImagingArchive.tcia_patients — Method

tcia_patients(; collection, format = "csv")

Returns the patients in a given collection.

CancerImagingArchive.tcia_patients_by_modality — Method

tcia_patients(; collection, modality, format = "csv")

Returns the patients in a given collection and modality (both inputs required).

CancerImagingArchive.tcia_series — Method

tcia_series(; collection, bodypart, manufacturer, modality, model, patient, series, study, format = "csv")

Returns series information for a given collection, bodypart,manufactuer,modality, manufacturermodel, patient, SeriesInstanceUID series, or StudyInstanceUID study.

CancerImagingArchive.tcia_series_size — Method

tcia_series_size(; series, format = "csv")

Returns the total byte size and the number of objects in the given SeriesInstanceUID series.

CancerImagingArchive.tcia_single_image — Method

tcia_single_image(; series, sop, file)

Downloads a single DICOM image for the given SeriesInstanceUID series and SOPInstanceUID sop. The DICOM file is saved as file.

CancerImagingArchive.tcia_sop — Method

tcia_sop(; series, format = "csv")

Returns the SOPInstanceUIDs for the given SeriesInstanceUID series.

CancerImagingArchive.tcia_studies — Method

tcia_studies(; collection, patient, study, format = "csv")

Returns the patient studies for a given collection and/or patient and/or study.

CancerImagingArchive.remove_empty! — Method

remove_empty!(dictionary::Dict)

Removes dictionary keys with empty values. Used internally to remove empty queries.